feat: update delete writer channel command

+++ __main__.py
@@ -62,8 +62,8 @@ async def on_interaction(interaction: Interaction):
     if interaction.type == InteractionType.component:
         if interaction.message.id == int(os.getenv("PUBLIC_MESSAGE_ID")):
             if (
-                utils.get(interaction.guild.roles, id=int(os.getenv("VIEWER")))
-                in interaction.user.roles
+                    utils.get(interaction.guild.roles, id=int(os.getenv("VIEWER")))
+                    in interaction.user.roles
             ):
                 await interaction.response.send_message(
                     embed=Embed(
@@ -94,7 +94,7 @@ async def writerApply(interaction: Interaction, channelName: str):
     await interaction.response.defer()
 
     if await database["channel"].find_one(
-        {"authors": {"$in": [str(interaction.user.id)]}}
+            {"authors": {"$in": [str(interaction.user.id)]}}
     ):
         await interaction.edit_original_response(
             embed=Embed(
@@ -234,7 +234,7 @@ async def warn(interaction: Interaction, channel: TextChannel):
         embed=Embed(
             title="⚠️ 경고",
             description="장기간 미활동으로 24시간 후 채널 삭제합니다!\n그림 올리시면 보존되니 참고 바랍니다!\n"
-            "`이 뒤로는 적어도 7일에 한번씩은 활동 부탁드려요!`",
+                        "`이 뒤로는 적어도 7일에 한번씩은 활동 부탁드려요!`",
             color=Color.red(),
         ),
     )
@@ -262,21 +262,71 @@ async def deleteWriterChannel(interaction: Interaction, channel: TextChannel):
             ephemeral=True,
         )
         return
-    await channel.delete()
-    findData = await database["channel"].find_one({"_id": str(channel.id)})
-    await database["channel"].delete_one({"_id": str(channel.id)})
-    member = utils.get(interaction.guild.members, id=int(findData["authors"][0]))
-    await member.remove_roles(
-        utils.get(interaction.guild.roles, id=int(os.getenv("WRITER")))
-    )
+
+    # confirm interaction
+    class Confirm(ui.View):
+        def __init__(self):
+            super().__init__(timeout=60)
+            self.value = None
+
+        @ui.button(label="확인", style=ButtonStyle.green, emoji="✅")
+        async def confirm(self, _interaction: Interaction, button: ui.Button):
+            self.value = True
+
+        @ui.button(label="거부", style=ButtonStyle.red, emoji="⛔")
+        async def cancel(self, _interaction: Interaction, button: ui.Button):
+            self.value = False
+
+    view = Confirm()
     await interaction.response.send_message(
         embed=Embed(
-            title="삭제",
-            description=f"정상적으로 삭제했습니다.",
-            color=Color.green(),
-        )
+            title="⚠️ Warning",
+            description=f"정말로 다음 채널을 삭제할까요?\n채널 : {channel.mention}",
+            color=Color.orange(),
+        ),
+        view=view,
     )
-    return
+    while True:
+        if view.timeout <= 0:
+            await interaction.edit_original_response(
+                embed=Embed(
+                    title="⚠️ Warning",
+                    description="시간이 초과되었습니다.",
+                    color=Color.red(),
+                )
+            )
+            return
+        await asyncio.sleep(1)
+        view.timeout -= 1
+        if view.value is not None and view.value is True:
+            await channel.delete()
+            findData = await database["channel"].find_one({"_id": str(channel.id)})
+            await database["channel"].delete_one({"_id": str(channel.id)})
+            member = utils.get(interaction.guild.members, id=int(findData["authors"][0]))
+            for role in [int(os.getenv("WRITER")), 982907252103086191, 982907265101201408, 982907268389535745]:
+                await member.remove_roles(
+                    utils.get(interaction.guild.roles, id=int(role))
+                )
+            await member.add_roles(
+                utils.get(interaction.guild.roles, id=int(os.getenv("VIEWER")))
+            )
+            await interaction.edit_original_response(
+                embed=Embed(
+                    title="✅ Success",
+                    description=f"정상적으로 삭제했습니다.",
+                    color=Color.green(),
+                )
+            )
+            return
+        elif view.value is not None and view.value is False:
+            await interaction.edit_original_response(
+                embed=Embed(
+                    title="⚠️ Warning",
+                    description="삭제를 취소했습니다.",
+                    color=Color.red(),
+                )
+            )
+            return
 
 
 @client.tree.command(
@@ -284,8 +334,8 @@ async def deleteWriterChannel(interaction: Interaction, channel: TextChannel):
 )
 async def tracker(interaction: Interaction) -> None:
     if not (
-        utils.get(interaction.guild.roles, id=int(os.getenv("VJ")))
-        in interaction.user.roles
+            utils.get(interaction.guild.roles, id=int(os.getenv("VJ")))
+            in interaction.user.roles
     ):
         await interaction.response.send_message(
             embed=Embed(
@@ -300,7 +350,7 @@ async def tracker(interaction: Interaction) -> None:
     trackedChannels: List[Tuple[TextChannel, datetime, int]] = []
     channelCount = 0
     for category in list(
-        filter(lambda x: x.name == "🎨【 작가채널 】", interaction.guild.categories)
+            filter(lambda x: x.name == "🎨【 작가채널 】", interaction.guild.categories)
     ):
         for channel in category.channels:
             if not channel.name.endswith("작가"):
@@ -316,7 +366,7 @@ async def tracker(interaction: Interaction) -> None:
                     tzinfo=timezone("Asia/Seoul")
                 )
             if (lastSendTime + timedelta(days=7)) < datetime.now(
-                tz=timezone("Asia/Seoul")
+                    tz=timezone("Asia/Seoul")
             ):
                 try:
                     trackedChannels.append(
@@ -351,8 +401,8 @@ async def tracker(interaction: Interaction) -> None:
 @client.tree.command(name="새로고침", description="( VJ ONLY ) 작가채널을 데이터베이스에 새로고침합니다.")
 async def refresh(interaction: Interaction) -> None:
     if not (
-        utils.get(interaction.guild.roles, id=int(os.getenv("VJ")))
-        in interaction.user.roles
+            utils.get(interaction.guild.roles, id=int(os.getenv("VJ")))
+            in interaction.user.roles
     ):
         await interaction.response.send_message(
             embed=Embed(
@@ -366,13 +416,13 @@ async def refresh(interaction: Interaction) -> None:
     await interaction.response.defer()
     abnormalChannels: List[TextChannel] = []
     for category in list(
-        filter(lambda x: x.name == "🎨【 작가채널 】", interaction.guild.categories)
+            filter(lambda x: x.name == "🎨【 작가채널 】", interaction.guild.categories)
     ):
         for channel in category.channels:
             if str(channel.topic) == "":
                 if (
-                    await database["channel"].find_one({"channel": str(channel.id)})
-                    is None
+                        await database["channel"].find_one({"channel": str(channel.id)})
+                        is None
                 ):
                     if channel.name.endswith("작가"):
                         abnormalChannels.append(channel)
@@ -393,7 +443,7 @@ async def refresh(interaction: Interaction) -> None:
             embed=Embed(
                 title="작가채널 새로고침",
                 description=f"정상적으로 작가채널을 새로고침 하였습니다.\n하지만, **{len(abnormalChannels)}** 개의 채널이\n"
-                f"비정상적으로 작동하고 있습니다.\n\n{', '.join([channel.mention for channel in abnormalChannels])}",
+                            f"비정상적으로 작동하고 있습니다.\n\n{', '.join([channel.mention for channel in abnormalChannels])}",
                 color=Color.red(),
             )
         )
